When zooming in, center cursor on screen
Until now, when zooming in, the zoom location was interpolated from the latest one, using the `follow` function alone to adjust the crop so that the cursor stays inside the active border. But, when not tracking the cursor, that meant a static zoom with the cursor offset to the sides, due to minimal necessary movement for it to be inside the active border.

Now, when zooming in, we set the initial zoom location to center the cursor in the window. So, whether tracking it or not, the initial transition will leave the cursor centralized on the screen (clamped, of course, to the configured zoom bounds and active border).

Besides ending up with the cursor centered, it also eliminates a side-travel effect that was noticeable when zooming in to a completely different location than the last zoomed-in one, cause by interpolation all the way from it.